fix multi ui perm conf redirect; consolidate perm selectors
