+++ ui/app/selectors/permissions.js
@@ -249,3 +249,32 @@ export function getPermissionsForActiveTab (state) {
     }
   })
 }
+
+export function getLastConnectedInfo (state) {
+  const { permissionsHistory = {} } = state.metamask
+  return Object.keys(permissionsHistory).reduce((acc, origin) => {
+    const ethAccountsHistory = JSON.parse(JSON.stringify(permissionsHistory[origin]['eth_accounts']))
+    return {
+      ...acc,
+      [origin]: ethAccountsHistory,
+    }
+  }, {})
+}
+
+export function getPermissionsRequests (state) {
+  return state.metamask.permissionsRequests || []
+}
+
+export function getPermissionsRequestCount (state) {
+  const permissionsRequests = getPermissionsRequests(state)
+  return permissionsRequests.length
+}
+
+export function getFirstPermissionRequest (state) {
+  const requests = getPermissionsRequests(state)
+  return requests && requests[0] ? requests[0] : null
+}
+
+export function hasPermissionRequests (state) {
+  return Boolean(getFirstPermissionRequest(state))
+}
